Un depreciated setting the username in MGTwitterEngine.
This is to clean up the method definitions for Lists. A number of the calls require the user to the
authorised user (like list creation) which means it doesn't make sense to be able to specify a user.

This may change in the future should Twitter allow people to collaborate on lists.